+++ stores/useSystemStore.js
@@ -32,6 +32,7 @@ export const useSystemStore = defineStore("system", () => {
     return (
       systemInstance.isMaintenance ||
       companyStore.company.maintenanceMode ||
+      companyStore.company.hasCompanyConfigurationReadError ||
       false
     );
   });